migrations bug fixed - migrations compeleted

+++ database/migrations/2021_12_13_144424_create_user_addresses_table.php
@@ -27,8 +27,8 @@ class CreateUserAddressesTable extends Migration
             $table->string("longitude")->nullable();
             $table->string("latitude")->nullable();
 
-            $table->unsignedBigInteger("province_id");
-            $table->unsignedBigInteger("city_id");
+            $table->unsignedInteger("province_id");
+            $table->unsignedInteger("city_id");
 
             $table->timestamps();
         });
